requirements.txt
tensorflow>=1.12
pillow
matplotlib
